requirements.txt
Django==1.4.3
django-extensions==1.0.3
Fabric==1.5.3
Werkzeug==0.8.3
py-bcrypt==0.2
South==0.7.6
django-tastypie==0.9.11
django-compressor==1.2